--- soojoob/src/main/java/freesia/soojoob/article/db/repository/ArticleRepository.java
@@ -1,4 +1,0 @@
-package freesia.soojoob.article.db.repository;
-
-public interface ArticleRepository {
-}
