+++ soojoob/src/main/java/freesia/soojoob/article/dto/request/ArticleUpdatePostReq.java
@@ -1,0 +1,4 @@
+package freesia.soojoob.article.dto.request;
+
+public class ArticleUpdatePostReq {
+}
