+++ soojoob/src/main/java/freesia/soojoob/article/dto/response/ArticleGetRes.java
@@ -1,0 +1,4 @@
+package freesia.soojoob.article.dto.response;
+
+public class ArticleGetRes {
+}
