+++ soojoob/src/main/java/freesia/soojoob/article/repository/ArticleRepository.java
@@ -1,0 +1,4 @@
+package freesia.soojoob.article.repository;
+
+public interface ArticleRepository {
+}
