--- soojoob/src/main/java/freesia/soojoob/article/request/ArticleRegisterPostReq.java
@@ -1,4 +1,0 @@
-package freesia.soojoob.article.request;
-
-public class ArticleRegisterPostReq {
-}
