--- soojoob/src/main/java/freesia/soojoob/article/request/ArticleUpdatePostReq.java
@@ -1,4 +1,0 @@
-package freesia.soojoob.article.request;
-
-public class ArticleUpdatePostReq {
-}
