--- soojoob/src/main/java/freesia/soojoob/article/response/ArticleGetRes.java
@@ -1,4 +1,0 @@
-package freesia.soojoob.article.response;
-
-public class ArticleGetRes {
-}
